Fix warnings in tutorials
After enabling extended warnings in the previous commits, the tutorials also emitted new warnings. This commit fixes them, there.

+++ tutorials/20-core/semaphore-worm-reenter.cpp
@@ -49,7 +49,7 @@ class worm : public StaticThread<> {
             last_x         = x;
             last_y         = y;
             printProtector.enter();
-            PRINTF(SCREEN_MOVE_CURSOR_TO, y, x);
+            PRINTF(SCREEN_MOVE_CURSOR_TO, static_cast<int>(y), static_cast<int>(x));
             if(x == RIGHT_LIMIT + 1)
                 PRINTF(":");
             else
@@ -83,7 +83,7 @@ class worm : public StaticThread<> {
             printProtector.enter();
             printProtector.enter();
             printProtector.enter();
-            PRINTF(SCREEN_MOVE_CURSOR_TO, y, x);
+            PRINTF(SCREEN_MOVE_CURSOR_TO, static_cast<int>(y), static_cast<int>(x));
             PRINTF("%c", letter);
             PRINTF(SCREEN_MOVE_CURSOR_TO, 0, 0);
             printProtector.leave();
